Final modifications to project structure and some comments
