+++ Comparison-shopping-engine/CoreEntities/Item.cs
@@ -6,8 +6,14 @@ using System.Threading.Tasks;
 
 namespace Comparison_shopping_engine
 {
+    /// <summary>
+    /// The class, representing the core item entity.
+    /// </summary>
     public class Item : IEquatable<Item>
     {
+        //Indicates whether the Item has been saved to the storage or not
+        public bool Saved
+        { get; set; }
         public string Name
         { get; set; }
 
@@ -83,6 +89,16 @@ namespace Comparison_shopping_engine
                 && new ItemStoreComparer().Compare(this, other) == 0
                 && new ItemPriceComparer().Compare(this, other) == 0
                 && new ItemDateComparer().Compare(this, other) == 0);
+        }
+
+        /// <summary>
+        /// Checks whether the data contained is valid. Used to determined whether the Item object
+        /// is ready to be saved to storage or not.
+        /// </summary>
+        /// <returns>The <see langword="bool"/> result, indicating whether the data is valid or not</returns>
+        public bool isComplete()
+        {
+            return (Store != "MissingSt" && Store != "") && (Price != 0);
         }
     }
 }
